Rename CIFAR-100 baselines folder to follow conventions from other directories.
PiperOrigin-RevId: 381066103
