+++ tensorflow_federated/python/simulation/baselines/cifar100/BUILD
@@ -14,15 +14,15 @@ package(default_visibility = [":default_visibility"])
 licenses(["notice"])
 
 py_library(
-    name = "cifar",
+    name = "cifar100",
     srcs = ["__init__.py"],
     srcs_version = "PY3",
     visibility = ["//tensorflow_federated/python/simulation/baselines:__pkg__"],
 )
 
 py_library(
-    name = "cifar_preprocessing",
-    srcs = ["cifar_preprocessing.py"],
+    name = "image_classification_preprocessing",
+    srcs = ["image_classification_preprocessing.py"],
     srcs_version = "PY3",
     deps = [
         "//tensorflow_federated/python/core/api:computation_base",
@@ -33,24 +33,24 @@ py_library(
 )
 
 py_test(
-    name = "cifar_preprocessing_test",
+    name = "image_classification_preprocessing_test",
     size = "medium",
-    srcs = ["cifar_preprocessing_test.py"],
+    srcs = ["image_classification_preprocessing_test.py"],
     python_version = "PY3",
     srcs_version = "PY3",
     deps = [
-        ":cifar_preprocessing",
+        ":image_classification_preprocessing",
         "//tensorflow_federated/python/core/backends/native:execution_contexts",
         "//tensorflow_federated/python/simulation/baselines:client_spec",
     ],
 )
 
 py_library(
-    name = "cifar100_tasks",
-    srcs = ["cifar100_tasks.py"],
+    name = "image_classification_tasks",
+    srcs = ["image_classification_tasks.py"],
     srcs_version = "PY3",
     deps = [
-        ":cifar_preprocessing",
+        ":image_classification_preprocessing",
         ":resnet_models",
         "//tensorflow_federated/python/learning:keras_utils",
         "//tensorflow_federated/python/learning:model",
@@ -62,12 +62,12 @@ py_library(
 )
 
 py_test(
-    name = "cifar100_tasks_test",
-    srcs = ["cifar100_tasks_test.py"],
+    name = "image_classification_tasks_test",
+    srcs = ["image_classification_tasks_test.py"],
     python_version = "PY3",
     srcs_version = "PY3",
     deps = [
-        ":cifar100_tasks",
+        ":image_classification_tasks",
         "//tensorflow_federated/python/core/backends/native:execution_contexts",
         "//tensorflow_federated/python/simulation/baselines:baseline_task",
         "//tensorflow_federated/python/simulation/baselines:client_spec",
