+++ tensorflow_federated/python/simulation/baselines/cifar100/image_classification_preprocessing.py
